+++ lib/change_QR/change_QR.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:provider/provider.dart';
 
 import '../edit_institute/edit_institute.dart';
 import '../setting.dart';
