+++ lib/mypage/my_page.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:provider/provider.dart';
 
 import '../edit_profile/edit_proflie_page.dart';
 import '../setting.dart';
